+++ packages/contents/src/index.ts
@@ -6,4 +6,4 @@ export * from './drivefs';
 export * from './tokens';
 export * from './broadcast';
 export * from './emscripten';
-export * from './drivecontent';
+export * from './drivecontents';
